sql_conn/requirements.txt
colorama==0.4.6
psycopg2-binary==2.9.9
colorama==0.4.6
python-dotenv==1.0.1
openpyxl==3.1.2